Team1 to DEV

+++ app/src/main/java/studentcompany/sportgest/daos/Squad_Call_DAO.java
@@ -12,6 +12,7 @@ import studentcompany.sportgest.daos.db.MyDB;
 import studentcompany.sportgest.daos.exceptions.GenericDAOException;
 import studentcompany.sportgest.domains.Game;
 import studentcompany.sportgest.domains.Player;
+import studentcompany.sportgest.domains.Team;
 
 public  class Squad_Call_DAO extends GenericPairDAO<Player,Game> implements IGenericPairDAO<Player,Game> {
     //Database name
@@ -25,13 +26,13 @@ public  class Squad_Call_DAO extends GenericPairDAO<Player,Game> implements IGen
     public static final String TABLE_NAME         = "SQUAD_CALL";
 
     //Table columns
-    public static final String COLUMN_GAME_ID = "GAME_ID";
     public static final String COLUMN_PLAYER_ID = "PLAYER_ID";
+    public static final String COLUMN_GAME_ID = "GAME_ID";
 
     //Create table
     public static final String CREATE_TABLE = "CREATE TABLE " + TABLE_NAME + " (" +
-            COLUMN_GAME_ID + " INTEGER NOT NULL, " +
             COLUMN_PLAYER_ID + " INTEGER NOT NULL, " +
+            COLUMN_GAME_ID + " INTEGER NOT NULL, " +
             "PRIMARY KEY (" + COLUMN_GAME_ID + ", " + COLUMN_PLAYER_ID + "), " +
             "FOREIGN KEY(" + COLUMN_GAME_ID + ") REFERENCES " + Game_DAO.TABLE_NAME + "(" + Game_DAO.COLUMN_ID + "), " +
             "FOREIGN KEY(" + COLUMN_PLAYER_ID + ") REFERENCES " + Player_DAO.TABLE_NAME + "(" + Player_DAO.COLUMN_ID + "));";
@@ -41,8 +42,8 @@ public  class Squad_Call_DAO extends GenericPairDAO<Player,Game> implements IGen
 
     public Squad_Call_DAO(Context context) {
         this.db = MyDB.getInstance(context).db;
-        this.game_dao = new Game_DAO(context);
         this.player_dao = new Player_DAO(context);
+        this.game_dao = new Game_DAO(context);
     }
 
     @Override
@@ -58,8 +59,8 @@ public  class Squad_Call_DAO extends GenericPairDAO<Player,Game> implements IGen
 
         //Parse data
         while(res.isAfterLast() == false) {
-            gameId = res.getLong(res.getColumnIndex(COLUMN_GAME_ID));
             playerId = res.getLong(res.getColumnIndex(COLUMN_PLAYER_ID));
+            gameId = res.getLong(res.getColumnIndex(COLUMN_GAME_ID));
             resList.add(
                     new Pair<>(
                             player_dao.getById(playerId),
@@ -112,6 +113,11 @@ public  class Squad_Call_DAO extends GenericPairDAO<Player,Game> implements IGen
 
     @Override
     public long insert(Pair<Player, Game> object) throws GenericDAOException {
+        if (object == null)
+            return -1;
+
+        if (object.getFirst() == null || object.getSecond() == null)
+            return -1;
         ContentValues contentValues = new ContentValues();
         contentValues.put(COLUMN_PLAYER_ID, object.getFirst().getId());
         contentValues.put(COLUMN_GAME_ID, object.getSecond().getId());
@@ -121,14 +127,67 @@ public  class Squad_Call_DAO extends GenericPairDAO<Player,Game> implements IGen
 
     @Override
     public boolean delete(Pair<Player, Game> object) throws GenericDAOException {
-        int deletedCount = db.delete(TABLE_NAME,
+        if (object == null)
+            return false;
+
+        if (object.getFirst() == null || object.getSecond() == null)
+            return false;
+
+        return db.delete(TABLE_NAME,
                 COLUMN_PLAYER_ID + " = ? , " + COLUMN_GAME_ID + " = ? ",
-                new String[] { Long.toString(object.getFirst().getId()), Long.toString(object.getSecond().getId()) });
-        return true;
+                new String[] { Long.toString(object.getFirst().getId()), Long.toString(object.getSecond().getId()) })  > 0;
+
     }
 
     @Override
     public boolean exists(Pair<Player, Game> object) throws GenericDAOException {
-        return false;
+        if (object == null)
+            return false;
+
+        if (object.getFirst() == null || object.getSecond() == null)
+            return false;
+
+        StringBuilder statement = new StringBuilder("SELECT * FROM " + TABLE_NAME + " where ");
+        statement.append(COLUMN_PLAYER_ID).append("=").append(object.getFirst().getId());
+        statement.append(" AND ").append(COLUMN_GAME_ID).append("=").append(object.getSecond().getId());
+
+        Cursor res = db.rawQuery(statement.toString(), null);
+        return res.moveToFirst();
+    }
+
+    public List<Player> getPlayersBy_GameID(long gameId) throws GenericDAOException{
+
+        int number;
+        long id,team;
+        String nickname, name, photo;
+        ArrayList<Player> players = null;
+
+        String sql = "SELECT * FROM " + Player_DAO.TABLE_NAME + " AS PL, " + Squad_Call_DAO.TABLE_NAME
+                + " AS SQ WHERE SQ." + Squad_Call_DAO.COLUMN_GAME_ID + "=" + gameId
+                + " AND SQ." + Squad_Call_DAO.COLUMN_PLAYER_ID + "=PL." + Player_DAO.COLUMN_ID;
+
+        //Query
+        Cursor res = db.rawQuery(sql, null);
+        res.moveToFirst();
+
+        if(res.isAfterLast() == false){
+            players = new ArrayList<Player>();
+
+            //Parse data
+            while(res.isAfterLast() == false) {
+                id = res.getLong(res.getColumnIndex(Player_DAO.COLUMN_ID));
+                nickname = res.getString(res.getColumnIndex(Player_DAO.COLUMN_NICKNAME));
+                name = res.getString(res.getColumnIndex(Player_DAO.COLUMN_NAME));
+                photo= res.getString(res.getColumnIndex(Player_DAO.COLUMN_PHOTO));
+                number= res.getInt(res.getColumnIndex(Player_DAO.COLUMN_NUMBER));
+                team=res.getLong(res.getColumnIndex(Player_DAO.COLUMN_TEAM_ID));
+
+
+                players.add(new Player(id,nickname,name,"","","",0,0,"","",photo,"","",number, new Team(team),null));
+                res.moveToNext();
+            }
+        }
+
+        return players;
     }
 }
